# chore(deps): remove peerDependencies