--- a/package.json
+++ b/package.json
@@ -14,7 +14,7 @@
   "scripts": {
     "prepare": "husky && bun run fetch:specs",
     "prebuild": "bun run --shell bun dist",
-    "build": "bun build ./src/index.ts --outdir ./dist --target node && tsc --emitDeclarationOnly",
+    "build": "tsdown",
     "rebuild": "bun run fetch:specs && bun run build",
     "prepublishOnly": "bun run rebuild",
     "test:unit": "bun test src",
@@ -38,18 +38,20 @@
     "@types/bun": "^1.2.4",
     "@types/json-schema": "^7.0.15",
     "@types/node": "^22.13.5",
-    "openai": "^4.85.4",
+    "ai": "^4.1.46",
     "dotenv": "^16.3.1",
     "husky": "^9.1.7",
     "lint-staged": "^15.2.0",
     "mkdocs": "^0.0.1",
+    "openai": "^4.85.4",
     "openapi-types": "^12.1.3",
+    "publint": "^0.3.12",
+    "tsdown": "^0.12.8",
     "typescript": "^5.0.0",
-    "ai": "^4.1.46"
+    "unplugin-unused": "^0.5.1"
   },
   "peerDependencies": {
     "ai": "4.x",
-    "@ai-sdk/openai": "1.x",
     "openai": "4.x"
   },
   "repository": {