--- a/package.json
+++ b/package.json
@@ -41,6 +41,7 @@
     "@types/bun": "^1.2.4",
     "@types/json-schema": "^7.0.15",
     "@types/node": "^22.13.5",
+    "@typescript/native-preview": "^7.0.0-dev.20250623.1",
     "ai": "^4.1.46",
     "dotenv": "^16.3.1",
     "husky": "^9.1.7",
@@ -50,7 +51,6 @@
     "openapi-types": "^12.1.3",
     "publint": "^0.3.12",
     "tsdown": "^0.12.8",
-    "typescript": "^5.0.0",
     "unplugin-unused": "^0.5.1"
   },
   "peerDependencies": {