--- a/tsdown.config.ts
+++ b/tsdown.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'tsdown';
+
+export default defineConfig({
+	entry: [ "./src/index.ts" ],
+	outDir: 'dist',
+	format: 'esm',
+	clean: true,
+	sourcemap: true,
+	treeshake: true,
+	dts: {
+    resolve: [/^@types\//]
+  },
+	publint: true,
+	unused: true,
+	unbundle: true,
+ });