--- a/tsdown.config.ts
+++ b/tsdown.config.ts
@@ -1,23 +1,25 @@
 import { defineConfig } from 'tsdown';
+import { $ } from 'bun';
 
 export default defineConfig({
-	entry: [ "./src/index.ts" ],
-	outDir: 'dist',
-	format: 'esm',
-	clean: true,
-	sourcemap: true,
-	treeshake: true,
-	dts: {
+  entry: ['./src/index.ts'],
+  outDir: 'dist',
+  format: 'esm',
+  clean: true,
+  sourcemap: true,
+  treeshake: true,
+  dts: {
     tsgo: true,
-    resolve: [/^@types\//]
+    resolve: [/^@types\//],
   },
-	publint: true,
-	unused: true,
-	unbundle: true,
-	exports: true,
- });
-  hooks:{
+  publint: true,
+  unused: true,
+  unbundle: true,
+  exports: true,
+  hooks: {
     'build:done': async () => {
       // sourcemap files for generated code are not needed
-      await $`rm -rf ./dist/openapi/generated/*.map`
-    }
+      await $`rm -rf ./dist/openapi/generated/*.map`;
+    },
+  },
+});