## Summary

This PR refactors the build system and optimizes dependencies to improve bundle size, build performance, and maintainability.

## Changes

### Build System Migration

- Migrated from manual TypeScript compilation to tsdown for a more streamlined build process
- Removed rimraf dependency in favor of built-in clean functionality
- Added proper exports field in package.json for better module resolution
- Enabled tsgo for faster TypeScript declaration generation using @typescript/native-preview

### Dependency Optimization

- Moved ai and openai to peerDependencies to prevent duplicate installations and reduce bundle size for consumers
- Moved json-schema from devDependencies to dependencies as it's required at runtime
- Removed unnecessary rimraf dependency by using Bun's built-in shell commands

### Build Configuration

- Configured tsdown with tree-shaking, source maps, and publint validation
- Enabled unused export detection to keep the codebase lean
- Set up unbundled ESM output for better tree-shaking in consumer applications
- Integrated tsgo for significantly faster type declaration generation

## Performance Improvements

### Build Time Comparison:
- bun tsdown: 677.5 ms (min: 659.1 ms, max: 698.5 ms)
- bun run build (previous): 1.188 s (min: 1.141 s, max: 1.283 s)

![SCR-20250623-orkc](https://github.com/user-attachments/assets/3759b8cb-1373-4d12-b61c-268dafa37238)

### Build Size Comparions
- `main` - 5.9MB
- `this branch` - 2.1 MB even though we added sourcemap

## Rationale

### Why switch from bundled to unbundled output?
- Previously, bun build was bundling all dependencies into a single file, which resulted in a large bundle size
- This forced consumers to include all code even when they only needed specific modules
- The bundled approach also included dependencies that should have been external, unnecessarily increasing the package size

### Why tsdown with unbundle mode instead of Bun build?
- Unbundled output: tsdown's unbundle mode preserves the module structure, allowing consumers to import only specific modules they need rather than the
entire bundle
- Better tree-shaking: Unbundled libraries enable consumer bundlers to perform more effective tree-shaking and code splitting
- Performance: tsdown is powered by Rolldown (Rust-based bundler), providing extremely fast build times compared to traditional bundlers
- Library-specific optimizations: Unlike general-purpose bundlers, tsdown is specifically optimized for building libraries with features like automatic
exports field generation

### Why peerDependencies?
- Both ai and openai are typically already installed in projects using this library
- This prevents version conflicts and duplicate installations
- Significantly reduces the final bundle size for consumers
- Allows consumers to control the exact versions they want to use

### Bundle Size Impact
- Moving heavy dependencies to peerDependencies reduces the installed size
- Tree-shaking with tsdown ensures only used code is included
- Unbundled output allows better optimization in consumer bundlers
- No longer bundling all dependencies into the output significantly reduces the package size

### TypeScript Declaration Generation
- Integrated tsgo (TypeScript's native preview) for much faster type declaration generation
- This experimental TypeScript feature provides significant performance improvements over traditional tsc declaration generation

## Breaking Changes

None - This maintains the same API surface while improving the build and dependency management.